Use np.float32

+++ gen_chem_1D/gen_models/reinvent/tokenization.py
@@ -48,7 +48,7 @@ class Vocabulary(object):
 
     def encode(self, char_list):
         """Takes a list of characters (eg '[NH]') and encodes to array of indices"""
-        smiles_matrix = np.zeros(len(char_list), dtype=np.float)
+        smiles_matrix = np.zeros(len(char_list), dtype=np.float32)
         for i, char in enumerate(char_list):
             smiles_matrix[i] = self.vocab[char]
         return smiles_matrix
